+++ src/Plugin.php
@@ -635,6 +635,8 @@ class Plugin extends BasePlugin
         Event::on(UserElement::class, UserElement::EVENT_BEFORE_DELETE, [$this->getSubscriptions(), 'beforeDeleteUserHandler']);
         Event::on(UserElement::class, UserElement::EVENT_BEFORE_DELETE, [$this->getOrders(), 'beforeDeleteUserHandler']);
 
+        Event::on(Address::class, Address::EVENT_AFTER_SAVE, [$this->getOrders(), 'afterSaveAddressHandler']);
+
         Event::on(
             UserElement::class,
             UserElement::EVENT_DEFINE_BEHAVIORS,