+++ src/services/Customers.php
@@ -319,8 +319,11 @@ class Customers extends Component
         $user = Craft::$app->getUsers()->ensureUserByEmail($order->email);
 
         if (!$user->getIsCredentialed()) {
+            $billingAddress = $order->getBillingAddress();
+            $shippingAddress = $order->getShippingAddress();
+
             if (!$user->fullName) {
-                $user->fullName = $order->getBillingAddress()?->fullName ?? $order->getShippingAddress()?->fullName ?? '';
+                $user->fullName = $billingAddress?->fullName ?? $shippingAddress?->fullName ?? '';
             }
 
             $user->username = $order->email;
@@ -333,6 +336,45 @@ class Customers extends Component
 
                 if (!$emailSent) {
                     Craft::warning('"registerUserOnOrderComplete" used to create the user, but couldn’t send an activation email. Check your email settings.', __METHOD__);
+                }
+
+                if ($billingAddress || $shippingAddress) {
+                    $newAttributes = ['ownerId' => $user->id];
+
+                    // If there is only one address make sure we don't add duplicates to the user
+                    if ($order->hasMatchingAddresses()) {
+                        $newAttributes['title'] = Craft::t('commerce', 'Address');
+                        $shippingAddress = null;
+                    }
+
+                    // Copy addresses to user
+                    if ($billingAddress) {
+                        $newBillingAddress = Craft::$app->getElements()->duplicateElement($billingAddress, $newAttributes);
+
+                        /**
+                         * Because we are cloning from an order address the `CustomerAddressBehavior` hasn't been instantiated
+                         * therefore we are unable to simply set the `isPrimaryBilling` property when specifying the new attributes during duplication.
+                         */
+                        if (!$newBillingAddress->hasErrors()) {
+                            $this->savePrimaryBillingAddressId($user, $newBillingAddress->id);
+
+                            if ($order->hasMatchingAddresses()) {
+                                $this->savePrimaryShippingAddressId($user, $newBillingAddress->id);
+                            }
+                        }
+                    }
+
+                    if ($shippingAddress) {
+                        $newShippingAddress = Craft::$app->getElements()->duplicateElement($shippingAddress, $newAttributes);
+
+                        /**
+                         * Because we are cloning from an order address the `CustomerAddressBehavior` hasn't been instantiated
+                         * therefore we are unable to simply set the `isPrimaryShipping` property when specifying the new attributes during duplication.
+                         */
+                        if (!$newShippingAddress->hasErrors()) {
+                            $this->savePrimaryShippingAddressId($user, $newShippingAddress->id);
+                        }
+                    }
                 }
             } else {
                 $errors = $user->getErrors();